refactor/#6 [refactor] : Some > like 코루틴으로 로직 수정2
