+++ app/src/main/java/com/devnuts/ruflu/data/repository/SomeRepositoryImpl.kt
@@ -1,7 +1,9 @@
 package com.devnuts.ruflu.data.repository
 
-import com.devnuts.ruflu.data.api.response.card.UserModel
+import com.devnuts.ruflu.data.api.response.NetworkResponse
+import com.devnuts.ruflu.data.api.response.home.model.UserModel
 import com.devnuts.ruflu.data.source.SomeDataSource
+import com.devnuts.ruflu.domain.entities.UserEntity
 import com.devnuts.ruflu.domain.repository.SomeRepository
 import retrofit2.Call
 import javax.inject.Inject
@@ -9,11 +11,11 @@ import javax.inject.Inject
 class SomeRepositoryImpl @Inject constructor(
     private val someDataSource: SomeDataSource
 ) : SomeRepository {
-    override fun getLikeMeList(): Call<ArrayList<UserModel>> {
+    override suspend fun getLikeMeList(): Result<List<UserEntity>> {
         return someDataSource.getLikeMeList()
     }
 
-    override fun addUserInMyMatchList(userId: String): Call<String> {
+    override suspend fun addUserInMyMatchList(userId: String): Result<NetworkResponse> {
         return someDataSource.addUserInMyMatchList(userId)
     }
 }
